+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.5.1 - 2020/04/29
+- Added blend modes from pyora (credited in file docstring): GRAINEXTRACT,
+GRAINMERGE, DIVIDE, HUE, SATURATION, COLOUR, LUMINOSITY
+
 ## 2020.5 - 2020/04/28
 - Added LSR support
 - Added make.py